// apps/inatrace-fe/src/api/model/apiPaginatedResponseApiFacility.ts
export interface ApiPaginatedResponseApiFacility {
  /**
   * Response status. OK for successful reponses.
   */
  status: ApiPaginatedResponseApiFacility.StatusEnum;
  /**
   * Simple message to explain client developers the reason for error.
   */
  errorMessage?: string;
  data?: ApiPaginatedListApiFacility;
  /**
   * Optional details for unexpected error responses.
   */
  errorDetails?: string;
  validationErrorDetails?: ApiValidationErrorDetails;
}

/**
 * Namespace for property- and property-value-enumerations of ApiPaginatedResponseApiFacility.
 */
export namespace ApiPaginatedResponseApiFacility {
  /**
   * All properties of ApiPaginatedResponseApiFacility.
   */
  export enum Properties {
    /**
     * Response status. OK for successful reponses.
     */
    status = 'status',
    /**
     * Simple message to explain client developers the reason for error.
     */
    errorMessage = 'errorMessage',
    data = 'data',
    /**
     * Optional details for unexpected error responses.
     */
    errorDetails = 'errorDetails',
    validationErrorDetails = 'validationErrorDetails',
  }

  /**
   * All possible values of status.
   */
  export enum StatusEnum {
    OK = 'OK',
    ERROR = 'ERROR',
    REQUESTBODYERROR = 'REQUEST_BODY_ERROR',
    VALIDATIONERROR = 'VALIDATION_ERROR',
    TOOMANYREQUESTS = 'TOO_MANY_REQUESTS',
    UNAUTHORIZED = 'UNAUTHORIZED',
    AUTHERROR = 'AUTH_ERROR',
    UPSTREAMHTTPERROR = 'UPSTREAM_HTTP_ERROR',
    INVALIDREQUEST = 'INVALID_REQUEST',
    INVALIDOREXPIREDSTORAGEKEY = 'INVALID_OR_EXPIRED_STORAGE_KEY',
    NOTIMPLEMENTED = 'NOT_IMPLEMENTED',
    NOTFOUND = 'NOT_FOUND',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiPaginatedResponseApiFacility',
      vars: [
        {
          isReadOnly: false,
          isEnum: true,
          datatypeWithEnum: 'ApiPaginatedResponseApiFacility.StatusEnum',
          required: true,
          name: 'status',
          classname: 'ApiPaginatedResponseApiFacility',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'errorMessage',
          classname: 'ApiPaginatedResponseApiFacility',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiPaginatedListApiFacility.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'data',
          classname: 'ApiPaginatedResponseApiFacility',
          dataType: 'ApiPaginatedListApiFacility',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiPaginatedListApiFacility',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'errorDetails',
          classname: 'ApiPaginatedResponseApiFacility',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiValidationErrorDetails.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'validationErrorDetails',
          classname: 'ApiPaginatedResponseApiFacility',
          dataType: 'ApiValidationErrorDetails',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiValidationErrorDetails',
        },
      ],
      validators: {
        status: [['required']],
        errorMessage: [],
        data: [],
        errorDetails: [],
        validationErrorDetails: [],
      },
    };
  }

  // export const ApiPaginatedResponseApiFacilityValidationScheme = {
  //     validators: [],
  //     fields: {
  //               status: {
  //                   validators: []
  //               },
  //               errorMessage: {
  //                   validators: []
  //               },
  //               data: {
  //                   validators: []
  //               },
  //               errorDetails: {
  //                   validators: []
  //               },
  //               validationErrorDetails: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiPaginatedResponseApiFacility>;
}