// apps/inatrace-fe/src/api/model/apiPaginatedListApiFacility.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * INATrace Services API OpenAPI documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */

import { ApiFacility } from './apiFacility';

/**
 * Response body for successful responses.
 */

export interface ApiPaginatedListApiFacility {
  /**
   * Response items.
   */
  items?: Array<ApiFacility>;
  /**
   * Count of all items satisfying 'paginatable' request.
   */
  count?: number;
  /**
   * Offset got from request
   */
  offset?: number;
  /**
   * Limit got from request
   */
  limit?: number;
}

/**
 * Namespace for property- and property-value-enumerations of ApiPaginatedListApiFacility.
 */
export namespace ApiPaginatedListApiFacility {
  /**
   * All properties of ApiPaginatedListApiFacility.
   */
  export enum Properties {
    /**
     * Response items.
     */
    items = 'items',
    /**
     * Count of all items satisfying 'paginatable' request.
     */
    count = 'count',
    /**
     * Offset got from request
     */
    offset = 'offset',
    /**
     * Limit got from request
     */
    limit = 'limit',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiPaginatedListApiFacility',
      vars: [
        {
          metadata: ApiFacility.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'items',
          classname: 'ApiPaginatedListApiFacility',
          dataType: 'Array&lt;ApiFacility&gt;',
          isPrimitiveType: false,
          isListContainer: true,
          complexType: 'ApiFacility',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'count',
          classname: 'ApiPaginatedListApiFacility',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'offset',
          classname: 'ApiPaginatedListApiFacility',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'limit',
          classname: 'ApiPaginatedListApiFacility',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
      ],
      validators: {
        items: [],
        count: [],
        offset: [],
        limit: [],
      },
    };
  }

  // export const ApiPaginatedListApiFacilityValidationScheme = {
  //     validators: [],
  //     fields: {
  //               items: {
  //                   validators: []
  //               },
  //               count: {
  //                   validators: []
  //               },
  //               offset: {
  //                   validators: []
  //               },
  //               limit: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiPaginatedListApiFacility>;
}