// apps/inatrace-fe/src/api/model/apiPaginatedResponseApiProcessingAction.ts
export interface ApiPaginatedResponseApiProcessingAction {
  /**
   * Response status. OK for successful reponses.
   */
  status: ApiPaginatedResponseApiProcessingAction.StatusEnum;
  /**
   * Simple message to explain client developers the reason for error.
   */
  errorMessage?: string;
  data?: ApiPaginatedListApiProcessingAction;
  /**
   * Optional details for unexpected error responses.
   */
  errorDetails?: string;
  validationErrorDetails?: ApiValidationErrorDetails;
}

/**
 * Namespace for property- and property-value-enumerations of ApiPaginatedResponseApiProcessingAction.
 */
export namespace ApiPaginatedResponseApiProcessingAction {
  /**
   * All properties of ApiPaginatedResponseApiProcessingAction.
   */
  export enum Properties {
    /**
     * Response status. OK for successful reponses.
     */
    status = 'status',
    /**
     * Simple message to explain client developers the reason for error.
     */
    errorMessage = 'errorMessage',
    data = 'data',
    /**
     * Optional details for unexpected error responses.
     */
    errorDetails = 'errorDetails',
    validationErrorDetails = 'validationErrorDetails',
  }

  /**
   * All possible values of status.
   */
  export enum StatusEnum {
    OK = 'OK',
    ERROR = 'ERROR',
    REQUESTBODYERROR = 'REQUEST_BODY_ERROR',
    VALIDATIONERROR = 'VALIDATION_ERROR',
    TOOMANYREQUESTS = 'TOO_MANY_REQUESTS',
    UNAUTHORIZED = 'UNAUTHORIZED',
    AUTHERROR = 'AUTH_ERROR',
    UPSTREAMHTTPERROR = 'UPSTREAM_HTTP_ERROR',
    INVALIDREQUEST = 'INVALID_REQUEST',
    INVALIDOREXPIREDSTORAGEKEY = 'INVALID_OR_EXPIRED_STORAGE_KEY',
    NOTIMPLEMENTED = 'NOT_IMPLEMENTED',
    NOTFOUND = 'NOT_FOUND',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiPaginatedResponseApiProcessingAction',
      vars: [
        {
          isReadOnly: false,
          isEnum: true,
          datatypeWithEnum:
            'ApiPaginatedResponseApiProcessingAction.StatusEnum',
          required: true,
          name: 'status',
          classname: 'ApiPaginatedResponseApiProcessingAction',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'errorMessage',
          classname: 'ApiPaginatedResponseApiProcessingAction',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiPaginatedListApiProcessingAction.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'data',
          classname: 'ApiPaginatedResponseApiProcessingAction',
          dataType: 'ApiPaginatedListApiProcessingAction',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiPaginatedListApiProcessingAction',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'errorDetails',
          classname: 'ApiPaginatedResponseApiProcessingAction',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiValidationErrorDetails.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'validationErrorDetails',
          classname: 'ApiPaginatedResponseApiProcessingAction',
          dataType: 'ApiValidationErrorDetails',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiValidationErrorDetails',
        },
      ],
      validators: {
        status: [['required']],
        errorMessage: [],
        data: [],
        errorDetails: [],
        validationErrorDetails: [],
      },
    };
  }

  // export const ApiPaginatedResponseApiProcessingActionValidationScheme = {
  //     validators: [],
  //     fields: {
  //               status: {
  //                   validators: []
  //               },
  //               errorMessage: {
  //                   validators: []
  //               },
  //               data: {
  //                   validators: []
  //               },
  //               errorDetails: {
  //                   validators: []
  //               },
  //               validationErrorDetails: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiPaginatedResponseApiProcessingAction>;
}